+++ src/lib/app-provider/hooks/useBaseApiRoute.ts
@@ -6,6 +6,7 @@ export const useBaseApiRoute = (
     | "txs"
     | "balances"
     | "assets"
+    | "blocks"
     | "projects"
     | "contracts"
     | "codes"
@@ -30,11 +31,13 @@ export const useBaseApiRoute = (
 
   switch (type) {
     case "txs":
-      return `${api}/txs/${chain}/${currentChainId}`;
+      return `${api}/v1/${chain}/${currentChainId}/txs`;
     case "balances":
       return `${api}/v1/${chain}/${currentChainId}/balances`;
     case "assets":
       return `${api}/v1/${chain}/${currentChainId}/assets`;
+    case "blocks":
+      return `${api}/v1/${chain}/${currentChainId}/blocks`;
     case "projects":
       return `${api}/projects/${chain}/${currentChainId}`;
     case "contracts":